requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
numpy==2.1.3
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
XlsxWriter==3.2.0
